pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = "Django-based kanikervanaf website"
authors = ["Lars van Rhijn"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
black = "^19.10b0"
pydocstyle = "^5.0.2"
django-phonenumber-field = "^4.0.0"
phonenumbers = "^8.11.5"
requests = "^2.23.0"
Weasyprint = "^51"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
